// test/integration/test/ports/engines/queries/filters/composite/none/_fixture/generated/contract.d.ts
export type CodecTypes = MongoCodecTypes;
// ...
export type FieldOutputTypes = {
  readonly __unbound__: {
    readonly TestModel: {
      readonly _id: CodecTypes['mongo/int32@1']['output'];
      readonly field: CodecTypes['mongo/string@1']['output'] | null;
      readonly top_a: ReadonlyArray<CompositeAOutput>;
      readonly to_one_b: CompositeBOutput | null;
    };
  };
};
export type FieldInputTypes = {
  readonly __unbound__: {
    readonly TestModel: {
      readonly _id: CodecTypes['mongo/int32@1']['input'];
      readonly field: CodecTypes['mongo/string@1']['input'] | null;
      readonly top_a: ReadonlyArray<CompositeAInput>;
      readonly to_one_b: CompositeBInput | null;
    };
  };
};
export type TypeMaps = MongoTypeMaps<CodecTypes, FieldOutputTypes, FieldInputTypes>;
// ...
export type Contract = MongoContractWithTypeMaps<ContractBase, TypeMaps>;